py: fix data directory search algorithm
Fix a bug by which the executable directory would not get scanned for
the `chesemagna-data`.
realpath(__file__) will return the full path of the executable: use
dirname to find its directory and search there for `chesemagna-data`.

Note that this bug doesn't affect windows because the batch file always
start the program in its directory.